% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newObj_tlars.R
\name{newObj_tlars}
\alias{newObj_tlars}
\title{Creates a Terminated-LARS (tlars) object}
\usage{
newObj_tlars(
  lars_state,
  X,
  y,
  L_val,
  verbose = FALSE,
  intercept = FALSE,
  normalize = TRUE,
  type = "lar"
)
}
\arguments{
\item{lars_state}{Variables associated with previous stopping point (necessary to restart forward selection selection process exactly where it was previously terminated). The lars_state is only required when the object (or its pointer) of the class tlarsCpp is deleted or got lost in another R session (e.g., in parallel processing).}

\item{X}{Real valued Predictor matrix.}

\item{y}{Response vector.}

\item{L_val}{Number of knockoffs that are appended to the predictor matrix.}

\item{verbose}{If TRUE progress in computations is shown.}

\item{intercept}{If TRUE an intercept is included.}

\item{normalize}{If TRUE the predictors are standardized and the response is centered.}

\item{type}{'lar' for 'LARS' and 'lasso' for Lasso.}
}
\value{
Object of the class tlarsCpp.
}
\description{
Creates an object of the class tlarsCpp.
}
\examples{
data('Gauss_data')
X = Gauss_data$X
y = c(Gauss_data$y)
p = ncol(X)
n = nrow(X)
knock = matrix(stats::rnorm(n * p), nrow = n, ncol = p)
XK = cbind(X, knock)
larsObj = newObj_tlars(X = X, y = y, L_val = p)
larsObj
}
