% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rcpp_tlars_cpp.R
\name{plot.Rcpp_tlars_cpp}
\alias{plot.Rcpp_tlars_cpp}
\title{Plots the T-LARS solution path}
\usage{
\method{plot}{Rcpp_tlars_cpp}(
  tlars_model,
  include_knocks = TRUE,
  actions = TRUE,
  col_selected = "black",
  col_knocks = "red",
  xlab = "# Included knockoffs",
  ylab = "Coefficients",
  lty_selected = "solid",
  lty_knocks = "dashed",
  legend_pos = "topleft",
  ...
)
}
\arguments{
\item{tlars_model}{Object of the class tlars_cpp. See \link{tlars_model} for details.}

\item{include_knocks}{Logical. If TRUE solution paths of knockoffs are added to the plot.}

\item{actions}{Logical. If TRUE axis above plot with indices of added variables
(Knockoffs represented by 'K') along the solution path is added.}

\item{col_selected}{Color of lines corresponding to selected variables.}

\item{col_knocks}{Color of lines corresponding to included knockoffs.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-axis.}

\item{lty_selected}{Line type of lines corresponding to selected variables.
See \link{par} for more details.}

\item{lty_knocks}{Line type of lines corresponding to included knockoffs.
See \link{par} for more details.}

\item{legend_pos}{Legend position. See \link{xy.coords} for more details.}

\item{...}{Other parameter that control the appearance of the plot. See \link{plot} and \link{par}.}
}
\description{
Plots the T-LARS solution path stored in C++ objects of class tlars_cpp
(see \link{tlars_model} for details) if the object is created with type = "lar"
(no plot for type = "lasso").
}
\examples{
data('Gauss_data')
X = Gauss_data$X
y = drop(Gauss_data$y)
p = ncol(X)
n = nrow(X)
knocks = matrix(stats::rnorm(n * p), nrow = n, ncol = p)
XK = cbind(X, knocks)
mod_tlars = tlars_model(X = XK, y = y, num_knocks = ncol(knocks))
tlars(model = mod_tlars, T_stop = 3, early_stop = TRUE)
plot(mod_tlars)
}
\seealso{
\link{tlars_model}, \link{plot}, \link{par}, and \link{xy.coords}.
}
