% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rcpp_tlars_cpp.R
\name{print.Rcpp_tlars_cpp}
\alias{print.Rcpp_tlars_cpp}
\title{Prints a summary of the results stored in a C++ object of class tlars_cpp.}
\usage{
\method{print}{Rcpp_tlars_cpp}(tlars_model, ...)
}
\arguments{
\item{tlars_model}{Object of the class tlars_cpp. See \link{tlars_model} for details.}

\item{...}{Additional parameters to print. See \link{print} for details.}
}
\description{
Prints a summary of the results stored in a C++ object of class tlars_cpp
(see \link{tlars_model} for details), i.e., selected variables, computation time,
and number of included knockoffs.
}
\examples{
data('Gauss_data')
X = Gauss_data$X
y = drop(Gauss_data$y)
p = ncol(X)
n = nrow(X)
knocks = matrix(stats::rnorm(n * p), nrow = n, ncol = p)
XK = cbind(X, knocks)
mod_tlars = tlars_model(X = XK, y = y, num_knocks = ncol(knocks))
tlars(model = mod_tlars, T_stop = 3, early_stop = TRUE)
print(mod_tlars)
}
\seealso{
\link{tlars_model}.
}
