% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlars.R
\name{tlars}
\alias{tlars}
\title{Executes the Terminated-LARS (tlars) algorithm}
\usage{
tlars(obj, T_stop = 1, earlyStop = TRUE)
}
\arguments{
\item{obj}{Object of the class tlarsCpp.}

\item{T_stop}{Number of included knockoffs after which the random experiments (i.e., forward selection processes) are stopped.}

\item{earlyStop}{If TRUE, then the forward selection process is stopped after T_stop knockoffs have been included. Otherwise the entire solution path is computed.}
}
\description{
Modifies the generic tlarsCpp object by executing the tlars algorithm and including the results in the tlarsCpp object.
}
\examples{
data('Gauss_data')
X = Gauss_data$X
y = c(Gauss_data$y)
p = ncol(X)
n = nrow(X)
knock = matrix(stats::rnorm(n * p), nrow = n, ncol = p)
XK = cbind(X, knock)
larsObj = newObj_tlars(X = XK, y = y, L_val = p)
tlars(obj = larsObj, T_stop = 3, earlyStop = TRUE)
beta = larsObj$getLastBeta()
beta
}
