% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tlarsCpp.R
\docType{data}
\name{tlarsCpp}
\alias{tlarsCpp}
\alias{tlars-package}
\title{tlars}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
tlarsCpp
}
\description{
tlars package
}
\details{
Imports
}
\keyword{datasets}
